requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
faiss-cpu>=1.7.4
langchain>=0.1.20
langchain-groq>=0.1.3
langchain-community>=0.0.38
sentence-transformers>=2.7.0
streamlit>=1.33.0
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
groq>=0.5.0
peft>=0.10.0
transformers>=4.40.0
torch>=2.0.0
joblib>=1.3.0